thingy52: Add BLE beacon code for sharing IMU data
This code implements a BLE beacon topology for sharing pitch,
roll and yaw data. Each pitch roll and yaw is shared via a separate
beacon using extended advertising. The packet format follows iBeacon
with the Major and Minor bits storing the float value. The last UUID
bit stores the type.

I forgot to commit this code last week and am commiting it as a
reference. I am going to refactor it into a single beacon with
a custom format.