pi: Implement ADS-B packet reception
This commit enables ADS-B packets to be received. A child process is
used to execute dump1090. The ADS-B data is then read in a JSON format
from dump1090's included web-server. Packets are cached in a
dictionary and only updated when the data changes.

The stderr of dump1090 is routed to the main process's logger to
facilitate debugging.